Benchtool: Move observers to a test-lib target [DPP-979]
changelog_begin
changelog_end
