+++ ledger/ledger-api-bench-tool/src/test/suite/scala/com/daml/ledger/api/benchtool/metrics/DelayMetricSpec.scala
@@ -1,9 +1,8 @@
 // Copyright (c) 2022 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.ledger.api.benchtool
+package com.daml.ledger.api.benchtool.metrics
 
-import com.daml.ledger.api.benchtool.metrics.DelayMetric
 import com.daml.ledger.api.benchtool.metrics.DelayMetric._
 import com.google.protobuf.timestamp.Timestamp
 import org.scalatest.matchers.should.Matchers
