+++ ledger/ledger-api-bench-tool/src/test/suite/scala/com/daml/ledger/api/benchtool/metrics/LatencyMetricSpec.scala
@@ -1,9 +1,8 @@
 // Copyright (c) 2022 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.ledger.api.benchtool
+package com.daml.ledger.api.benchtool.metrics
 
-import com.daml.ledger.api.benchtool.metrics.LatencyMetric
 import com.daml.ledger.api.benchtool.metrics.LatencyMetric.MaxLatency
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpec
